% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.otu.table.R
\docType{data}
\name{throat.otu.table}
\alias{throat.otu.table}
\title{OTU count table from 16S sequencing of the throat microbiome samples}
\format{
A data frame with 60 observations on 856 variables.
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010)
Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\usage{
data("throat.otu.tab")
}
\description{
This data set contains 60 subjects with 28 smokers and 32 nonsmokers.
Microbiome data were collected from right and left nasopharynx and oropharynx
region to form an OTU table with 856 OTUs.
}
\examples{
data(throat.otu.table)
}
\references{
R package "GUniFrac"
}
\keyword{datasets}
