% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.otu.taxonomy.R
\docType{data}
\name{throat.otu.taxonomy}
\alias{throat.otu.taxonomy}
\title{Taxanomy info of OTUs of 16S sequencing of the throat microbiome samples}
\format{
A vector with 5683 elements
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010)
Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\usage{
data(throat.otu.taxonomy)
}
\description{
This data set contains taxonomy info of 5683 otus, 856 of which are otus in the throat.otu.table.
}
\examples{
data(throat.otu.taxonomy)
}
\keyword{datasets}
